# public/taller-api-swagger-docs.json
{"openapi":"3.0.0","paths":{"/api/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/v1/payments/initialize":{"post":{"operationId":"PaymentsController_initializePayment","parameters":[],"responses":{"200":{"description":"Payment initialized successfully","content":{"application/json":{"schema":{"example":{"transactionId":"TXN_1234567890_abc123","provider":"transbank","redirectUrl":"https://payment-provider.com/checkout","token":"token_123","message":"Payment initialized successfully"}}}}},"400":{"description":"Invalid payment data"},"402":{"description":"Payment provider error"}},"security":[{"bearer":[]}],"summary":"Initialize a new payment transaction","tags":["Payments"]}},"/api/v1/payments/confirm":{"post":{"operationId":"PaymentsController_confirmPayment","parameters":[],"responses":{"200":{"description":"Payment confirmed successfully"},"402":{"description":"Payment provider error"},"404":{"description":"Payment not found"}},"security":[{"bearer":[]}],"summary":"Confirm payment after customer completes the transaction","tags":["Payments"]}},"/api/v1/payments/status/{transactionId}":{"get":{"operationId":"PaymentsController_getPaymentStatus","parameters":[{"name":"provider","required":true,"in":"query","description":"Payment provider","schema":{"enum":["transbank","mercado_pago"],"type":"string"}},{"name":"transactionId","required":true,"in":"path","description":"Transaction ID","schema":{}}],"responses":{"200":{"description":"Payment status retrieved successfully"},"404":{"description":"Payment not found"}},"security":[{"bearer":[]}],"summary":"Get payment status by transaction ID","tags":["Payments"]}},"/api/v1/payments/refund/{transactionId}":{"post":{"operationId":"PaymentsController_refundPayment","parameters":[{"name":"transactionId","required":true,"in":"path","description":"Transaction ID","schema":{}}],"responses":{"200":{"description":"Payment refunded successfully"},"400":{"description":"Cannot refund this payment"},"404":{"description":"Payment not found"}},"security":[{"bearer":[]}],"summary":"Refund a captured payment","tags":["Payments"]}},"/api/v1/payments/webhook/callback":{"post":{"operationId":"PaymentsController_handleCallback","parameters":[],"responses":{"200":{"description":"Webhook processed successfully"},"400":{"description":"Invalid webhook data"}},"security":[{"bearer":[]}],"summary":"Receive payment provider callbacks/webhooks","tags":["Payments"]}},"/api/v1/payments/list":{"get":{"operationId":"PaymentsController_listPayments","parameters":[{"name":"limit","required":false,"in":"query","schema":{"example":10,"type":"number"}},{"name":"status","required":false,"in":"query","schema":{"enum":["pending","authorized","captured","declined","cancelled","refunded","error"],"type":"string"}}],"responses":{"200":{"description":"Payments retrieved successfully"}},"security":[{"bearer":[]}],"summary":"List payments with optional status filter","tags":["Payments"]}}},"info":{"title":"Taller API","description":"API documentation for Taller","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{}}}